requirements.txt
coloredlogs==10.0
path.py==12.0.1
requests==2.22.0
